+++ pyproject.toml
@@ -63,6 +63,10 @@ testpaths = ["tests"]
 python_files = ["test_*.py", "*_test.py"]
 python_classes = ["Test*"]
 python_functions = ["test_*"]
+pythonpath = [
+    "src",
+    "."
+]
 addopts = [
     "-v",
     "--tb=short",
