+++ src/common/config/__init__.py
@@ -1,0 +1,3 @@
+from .config import config
+
+__all__ = ["config"]
