+++ src/common/decorators/__init__.py
@@ -1,0 +1,3 @@
+from .log_decorator import log_node_entry
+
+__all__ = ["log_node_entry"]
