+++ src/meetingmuse/llm_models/hugging_face.py
@@ -1,6 +1,6 @@
 from langchain_huggingface import ChatHuggingFace, HuggingFaceEndpoint
 
-from common.config.config import config
+from common.config import config
 
 
 class HuggingFaceModel:
