+++ src/meetingmuse/nodes/human_interrupt_retry_node.py
@@ -3,7 +3,7 @@ from typing import Any
 from langchain_core.messages import AIMessage
 from langgraph.types import Command, interrupt
 
-from common.decorators.log_decorator import log_node_entry
+from common.decorators import log_node_entry
 from common.logger import Logger
 from meetingmuse.models.interrupts import InterruptInfo, InterruptType
 from meetingmuse.models.node import NodeName
