+++ src/server/services/server_lifecycle_manager.py
@@ -8,7 +8,7 @@ from typing import Any, Optional
 
 import uvicorn
 
-from common.config.config import config as env_config
+from common.config import config as env_config
 
 from .websocket_connection_service import WebSocketConnectionService
