+++ tests/common/utils/test_utils.py
@@ -2,9 +2,9 @@ import pytest
 from langchain_core.messages import AIMessage, HumanMessage
 from pydantic import ValidationError
 
+from common.utils import Utils
 from meetingmuse.models.meeting import MeetingFindings
 from meetingmuse.models.state import MeetingMuseBotState
-from meetingmuse.utils.utils import Utils
 
 
 class TestUtils:
